+++ mypyc/codegen/emitmodule.py
